code: add dynamic DOM fallback for none browser enviroments
Adds a DOM instance with window and document if the current
enviroment does not support it i.e (if its not a browser).

This will help validate some objects that are not supported
by node.js such as HTMLElement (https://developer.mozilla.o
rg/en-US/docs/Web/API/HTMLElement and etc.